changes to utils structure and corresponding inheritence
